+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
 
 - Add checkbox component
 - Add switch component
+- Add toggler component ([PR 14](https://github.com/input-output-hk/react-polymorph/pull/14))
 
 ## 0.2.7